moved 'ISettingsChangeListener' to global interfaces
